update upload image

+++ DemoApi/Services/UploadImage/UploadImageService.cs
@@ -1,20 +1,28 @@
-﻿using Microsoft.AspNetCore.Hosting;
+﻿using Dapper;
+using DemoApi.Context;
+using DemoApi.Models.Dtos.UserDtos;
+using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
+using Newtonsoft.Json;
 using System.Collections.Generic;
+using System.Data;
 using System.IO;
 using System.Linq;
 using System.Text.Json;
 using System.Threading.Tasks;
+using static System.Net.Mime.MediaTypeNames;
 
 namespace DemoApi.Services.UploadImage
 {
     public class UploadImageService: IUploadImageService
     {
         private readonly IWebHostEnvironment _env;
+        private readonly DapperConnection _dapperConnection;
 
-        public UploadImageService(IWebHostEnvironment env)
+        public UploadImageService(DapperConnection dapperConnection,IWebHostEnvironment env)
         {
-            _env = env;
+            _dapperConnection = dapperConnection;
+            _env = env ?? throw new ArgumentNullException(nameof(env));
         }
 
         public async Task<List<object>> UploadImagesAsync(List<IFormFile> files)
@@ -59,7 +67,7 @@ namespace DemoApi.Services.UploadImage
 
             return imageUrl;
         }
-        public bool DeleteImage(string imagePath)
+        public async Task<bool> DeleteImageAsync(string imagePath, int idImage)
         {
             if (string.IsNullOrEmpty(imagePath))
             {
@@ -75,11 +83,73 @@ namespace DemoApi.Services.UploadImage
 
             if (File.Exists(fullPath))
             {
-                File.Delete(fullPath);
-                return true; 
+                var deleteImage = "sp_productimage_delete";
+
+                using (var connection = _dapperConnection.GetConnection())
+                {
+                    await connection.OpenAsync();
+
+                    try
+                    {
+                        var Parameters = new DynamicParameters(new
+                        {
+                            p_ID = idImage
+                        });
+
+                        var rowsAffected = await connection.ExecuteAsync(
+                            deleteImage,
+                            Parameters,
+                            commandType: CommandType.StoredProcedure
+                        );
+
+                        if (rowsAffected > 0)
+                        {
+                            File.Delete(fullPath);
+                            return true;
+                        }
+                    }
+                    catch (Exception ex)
+                    {
+                        Console.WriteLine($"Failed to execute AddUser stored procedure: {ex.Message}");
+                        return false;
+                    }
+                }
+
             }
 
             return false; 
+        }
+
+        public async Task<int> GetIdImageAsync(string imagePath)
+        {
+            var queryGetId = "sp_productimage_getid";
+
+            using (var connection = _dapperConnection.GetConnection())
+            {
+                await connection.OpenAsync();
+
+                try
+                {
+                    
+                    var parametersGetId = new DynamicParameters();
+                    parametersGetId.Add("p_imageUrl",imagePath,DbType.String,ParameterDirection.Input);
+
+                    var idImage = await connection.QueryFirstOrDefaultAsync<int?>(queryGetId,parametersGetId,commandType: CommandType.StoredProcedure);
+
+                    if (idImage == null)
+                    {
+                        Console.WriteLine("Image ID not found in database.");
+                        return 0;
+                    }
+                    return (int)idImage;
+                    
+                }
+                catch (Exception ex)
+                {
+                    Console.WriteLine($"Failed to execute AddUser stored procedure: {ex.Message}");
+                    return 0;
+                }
+            }
         }
     }
 }
